Refactor AccountsList component
